Add Event class-based Serializer to serializers and events handler
